+++ src/scss/styles.scss
@@ -1,3 +1,3 @@
-@import './settings/vars';
-@import './settings/default';
-@import './settings/keyframes';
+@import './settings/vars.scss';
+@import './settings/default.scss';
+@import './settings/keyframes.scss';
