+++ src/sections/Footer/footer.scss
@@ -1,6 +1,4 @@
-body {
-  margin: 0;
-}
+@import '../../scss/styles.scss';
 
 footer {
   background: transparent url("../../images/bg-footer-h.jpg") no-repeat;
