+++ src/sections/allies/allies.scss
@@ -1,4 +1,4 @@
-@import '../../scss/settings/vars.scss';
+@import '../../scss/styles.scss';
 
 .allies {
   display: flex;
