+++ src/sections/copy/copy.scss
@@ -1,4 +1,4 @@
-@import '../../scss/settings/vars.scss';
+@import '../../scss/styles.scss';
 
 .copy__contact {
   color: $white;
