+++ src/sections/footer-title/footer-title.scss
@@ -1,4 +1,4 @@
-@import '../../scss/settings/vars.scss';
+@import '../../scss/styles.scss';
 
 .footer-title__concat { 
   color: $white;
