+++ src/sections/media/media.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from './node_modules/react';
 import './media.scss';
 import Image from '../../units/Image';
