+++ src/sections/media/media.scss
@@ -1,4 +1,4 @@
-@import '../../scss/settings/vars.scss';
+@import '../../scss/styles.scss';
 
 .media {
   padding-bottom: 5%;
